Introduce fuzzing for `lading_payload::block::Cache::fixed` (#727)
Introduce fuzzing for `lading_payload::block::Cache::fixed`

This commit introduces a new fuzzing harness for the above mentioned
function. We want to ensure that we are not crashing at runtime or consuming
more memory than allowed.

REF SMPTNG-79

+++ lading_payload/src/dogstatsd/metric.rs
@@ -33,7 +33,7 @@ impl MetricGenerator {
         name_length: ConfRange<u16>,
         multivalue_count: ConfRange<u16>,
         multivalue_pack_probability: f32,
-        metric_weights: &WeightedIndex<u8>,
+        metric_weights: &WeightedIndex<u16>,
         container_ids: Vec<String>,
         tagsets: common::tags::Tagsets,
         str_pool: &strings::Pool,
